w3c-reports/12092014/efl/runned-tests/html/syntax/parsing/template/clearing-the-stack-back-to-a-given-context/clearing-stack-back-to-a-table-row-context-expected.txt

CONSOLE MESSAGE: line 56: ReferenceError: Can't find variable: newHTMLDocument

Harness Error. harness_status.status = 1 , harness_status.message = ReferenceError: Can't find variable: newHTMLDocument
